ATTA/requirements.txt
pandas
rich
questionary
requests
pytest
python-dotenv